Optimization added caching layer
The way esx_billing is designed is flawed, every time a player accesses the billing menu, it makes a database call, so if a user spams the menu, it will actually hit the database quite a few times.

This updated fix this issue by using a caching layer, first time a user opens there billing menu the data will be fetched from the database, and cached to a table.

What’s changed?

esx_billing:getBills & esx_billing:getTargetBills both fetch data from the database, to get a players bills, both are now cached on first usage, so if a cop checks a players bills they are cached, so when the players checks there bills, it’s cached and don’t trigger a database call.

When a new bill is given, if there is a cache table active, it will add the new bill to the table, that is why I changed MySQL.Async.execute to MySQL.Async.insert, since it returns the id of the new bill given to a player.

When a bill Is payed it loop’s thru the entry’s in the cached table for the specific user, and finds an entry that matches the id of the bill, and removes it from the caching table.

On client drop, it removes their entry’s from the caching table.

Ill do some more changes to esx_billing, but this is first step of optimizing the resource, and avoiding it being able to hit MySQL every time the billing menu is requested.